// packages/dataframe/ts/verbs/utility/resample.types.d.ts
export type Frequency = `${number}S` | `${number}min` | `${number}H` | `${number}D` | `${number}W` | `${number}M` | `${number}Q` | `${number}Y` | number | {
    value: number;
    unit: "ms" | "s" | "min" | "h" | "d" | "w" | "M" | "Q" | "Y";
};